requirements.txt
bidict==0.22.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
python-dotenv==0.20.0
python-engineio==4.3.1
python-socketio==5.5.2
rel==0.4.7
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.1
